Allow HTTP client specification for `ListenBrainz`
This aligns the constructors to match the approach taken by the `Query`
class of the MusicBrainz library. Either you use the constructor where
you pass a client (in which case you are responsible for setting
whatever default user agent you want), or you use one of the other a
constructors to set up the user agent values to use with the
internally-created client(s).

The `[Default]ContactInfo` and `[Default]ProductInfo` properties have
been dropped in favour of a `[Default]UserAgent` property (a list of
product info header values).

There are also two additional methods, `ConfigureClient` and
`ConfigureClientCreation`, which allow exercising some control over
clients created on-demand by the class.