Add site-wide artist map support
This adds support for the `/1/stats/sitewide/artist-map` endpoint.

The `Countries` field has been moved from `IUserArtistMap` to a new,
separate `IArtistMap` interface; this is a breaking change.

The `forceRecalculation` parameter was dropped from the user-based
`GetArtistMapAsync()`; it is not really documented, nor does it seem to
have any effect.

The version number has been bumped to 5.0.0 (should have already been
done for earlier breaking changes).